refactor: `domain` 패키지 생성, 클래스 위치 이동
