+++ src/main/java/racingcar/RacingCarGameApplication.java
@@ -1,5 +1,6 @@
 package racingcar;
 
+import racingcar.domain.RacingGame;
 import racingcar.view.IOManager;
 
 import java.util.List;
