+++ src/main/java/racingcar/domain/MovementCondition.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.Random;
