--- src/test/java/racingcar/AllForwardMovementCondition.java
@@ -1,8 +1,0 @@
-package racingcar;
-
-class AllForwardMovementCondition extends MovementCondition {
-    @Override
-    public boolean satisfies() {
-        return true;
-    }
-}
