--- src/test/java/racingcar/AllStayMovementCondition.java
@@ -1,8 +1,0 @@
-package racingcar;
-
-class AllStayMovementCondition extends MovementCondition {
-    @Override
-    public boolean satisfies() {
-        return false;
-    }
-}
