+++ src/test/java/racingcar/domain/AllForwardMovementCondition.java
@@ -1,0 +1,8 @@
+package racingcar.domain;
+
+public class AllForwardMovementCondition extends MovementCondition {
+    @Override
+    public boolean satisfies() {
+        return true;
+    }
+}
