+++ src/test/java/racingcar/domain/AllStayMovementCondition.java
@@ -1,0 +1,8 @@
+package racingcar.domain;
+
+public class AllStayMovementCondition extends MovementCondition {
+    @Override
+    public boolean satisfies() {
+        return false;
+    }
+}
