+++ src/test/java/racingcar/domain/FirstCarForwardMovementCondition.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class FirstCarForwardMovementCondition extends MovementCondition {
