+++ src/test/java/racingcar/domain/SecondCarForwardMovementCondition.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class SecondCarForwardMovementCondition extends MovementCondition {
