+++ src/test/java/racingcar/view/KeyboardInputTest.java
@@ -1,32 +1,16 @@
 package racingcar.view;
 
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.PrintStream;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class KeyboardInputTest {
 
-    private final ByteArrayOutputStream outputStreamCaptor = new ByteArrayOutputStream();
-    private final PrintStream standardOut = System.out;
     private Input input;
-
-    @BeforeEach
-    void setUp() {
-        System.setOut(new PrintStream(outputStreamCaptor));
-    }
-
-    @AfterEach
-    void tearDown() {
-        System.setOut(standardOut);
-    }
 
     @DisplayName("comma를 포함한 자동차 이름들을 인자로 받아 리스트 형태로 출력")
     @Test
@@ -39,7 +23,6 @@ public class KeyboardInputTest {
         List<String> carNames = input.getCarNames();
 
         // then
-        assertThat(outputStreamCaptor.toString().trim()).isEqualTo("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
         assertThat(carNames.get(0)).isEqualTo("abc");
         assertThat(carNames.get(1)).isEqualTo("def");
         assertThat(carNames.get(2)).isEqualTo("ghi");
@@ -57,7 +40,6 @@ public class KeyboardInputTest {
 
         // then
         assertThat(result).isEqualTo(Integer.parseInt(inputString));
-        assertThat(outputStreamCaptor.toString().trim()).isEqualTo("시도할 횟수는 몇 회인가요?");
     }
 
     private void provideInput(final String inputString) {
